lib:printf now supports 64bit ints

+++ lib/include/stdlib.h
@@ -8,7 +8,7 @@
 #ifndef __STDLIB_H__
 #define __STDLIB_H__
 
-char* itoa(int num);
+char* itoa(long long num);
 int stoi(char* str);
 
 #endif
