+++ lib/src/stdio.c
@@ -60,7 +60,7 @@ void printf(char* fmt, ...){
 					uart_write(str);
 					break;
 				case 'd':
-				    val = va_arg(ap, int);
+				    val = va_arg(ap, long long);
 					uart_write(itoa(val));
 					break;
 				case 'c':
